requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
# psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
requests==2.31.0
python-json-logger==2.0.7
# Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
